feat(hm): replace rofi with walker for app launcher
This commit replaces rofi with walker as the primary application
launcher in hyprland.

The new `walker.nix` module configures walker with several plugins,
including application search, calculator, clipboard manager, dmenu, and
emojis.

The hyprland configuration is updated to use walker, and the rofi
configuration is removed from the linux-desktop module.